+++ NAMESPACE
@@ -75,7 +75,6 @@ export(smr)
 export(smr_demo)
 export(sparseRowVariances)
 export(svd_lsi)
-export(testBackend)
 export(test_backends)
 export(tf_idf_transform)
 export(tf_idf_transform_v2)
