+++ R/RcppExports.R
@@ -42,11 +42,6 @@ naive_bayes_demo <- function(perc = 80L, verbose = TRUE) {
 }
 
 #' @export
-test_backends <- function() {
-    .Call('_viewmastR_test_backends', PACKAGE = 'viewmastR')
-}
-
-#' @export
 af_nn <- function(train_feats, test_feats, train_target, test_target, num_classes, layers, query_feats, relu_activation = FALSE, device = 0L, dts = "f32", learning_rate = 2.0, max_epochs = 250L, batch_size = 100L, max_error = 0.5, verbose = TRUE, benchmark = FALSE) {
     .Call('_viewmastR_af_nn', PACKAGE = 'viewmastR', train_feats, test_feats, train_target, test_target, num_classes, layers, query_feats, relu_activation, device, dts, learning_rate, max_epochs, batch_size, max_error, verbose, benchmark)
 }
@@ -81,11 +76,6 @@ computeSparseRowVariances <- function(j, val, rm, n) {
 }
 
 #' @export
-testBackend <- function() {
-    invisible(.Call('_viewmastR_testBackend', PACKAGE = 'viewmastR'))
-}
-
-#' @export
 get_sigmoid <- function(input) {
     .Call('_viewmastR_get_sigmoid', PACKAGE = 'viewmastR', input)
 }
@@ -98,5 +88,10 @@ get_relu <- function(input) {
 #' @export
 get_mnist <- function(perc = 80L, verbose = TRUE) {
     .Call('_viewmastR_get_mnist', PACKAGE = 'viewmastR', perc, verbose)
+}
+
+#' @export
+test_backends <- function() {
+    .Call('_viewmastR_test_backends', PACKAGE = 'viewmastR')
 }
