+++ src/nn.cpp
@@ -26,44 +26,6 @@
 
 using namespace af;
 
-static void testBackend()
-{
-  info();
-  af_print(randu(5, 4));
-}
-
-//' @export
-// [[Rcpp::export]]
-int test_backends()
-  {
-    try {
-      fprintf(stderr,"Trying CPU Backend\n");
-      setBackend(AF_BACKEND_CPU);
-      testBackend();
-    } catch (exception& e) {
-      fprintf(stderr,"Caught exception when trying CPU backend\n");
-      fprintf(stderr, "%s\n", e.what());
-    }
-    // try {
-    //   fprintf(stderr,"Trying CUDA Backend\n");
-    //   af::setBackend(AF_BACKEND_CUDA);
-    //   testBackend();
-    // } catch (af::exception& e) {
-    //   fprintf(stderr,"Caught exception when trying CUDA backend\n");
-    //   fprintf(stderr, "%s\n", e.what());
-    // }
-    try {
-      fprintf(stderr, "Trying OpenCL Backend\n");
-      setBackend(AF_BACKEND_OPENCL);
-      testBackend();
-    } catch (exception& e) {
-      fprintf(stderr,"Caught exception when trying OpenCL backend\n");
-      fprintf(stderr, "%s\n", e.what());
-    }
-    return 0;
-  }
-
-
 using std::vector;
